requirements.txt
langchain
langchain-openai
openai
requests
python-dotenv
